server/auth/requirements.txt
fastapi
fastapi-sso
uvicorn
sqlalchemy
pymysql
pydantic
pyjwt
boto3
passlib[bcrypt]
python-dotenv
requests